feat(models): add SumRequest and SumResponse
