+++ internal/domain/models/models.go
@@ -1,0 +1,10 @@
+package models
+
+type SumRequest struct {
+	FirstNumber  int `json:"first_number"`
+	SecondNumber int `json:"second_number"`
+}
+
+type SumResponse struct {
+	Result int `json:"result"`
+}
